requirements.txt
Flask==3.1.3
torch==2.9.0
torchvision==0.24.0
numpy
opencv-python-headless
Pillow
bcrypt